[chore] moved summoner-window files to analyst-drawer module
