+++ src/app/analyst-drawer-app/analyst-drawer-layout/analyst-drawer-layout.component.html
@@ -2,8 +2,8 @@
   class="analyst-drawer-layout"
   fxLayout="row"
   fxLayoutGap="10px">
-  <div class="analyst-content" cdkDrag>
+  <summoner-window-frame class="analyst-content" cdkDrag>
     Dragable Analyst Desk Make Sure You Make This Content Absolute
-  </div>
+  </summoner-window-frame>
 </div>
