+++ src/app/analyst-drawer-app/analyst-drawer-model.ts
@@ -1,0 +1,20 @@
+export enum RunePath {
+    Domination,
+    Precision,
+    Inspiration,
+    Sorcery,
+    Resolve
+}
+
+export class Rune {
+    name: string;
+    description: string;
+    path: RunePath;
+}
+
+export class BuildItem {
+    name: string;
+    icon: string;
+    description: string;
+    cost: number;
+}
