+++ src/app/analyst-drawer-app/summoner-window-content/summoner-window-content.stories.ts
