+++ src/app/analyst-drawer-app/summoner-window-frame/summoner-window-frame.component.scss
