+++ src/app/analyst-drawer-app/summoner-window-last-build/summoner-window-last-build.component.scss
