+++ src/app/analyst-drawer-app/summoner-window-last-build/summoner-window-last-build.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { BuildItem } from '../match-viewer-model';
+import { BuildItem } from '../analyst-drawer-model';
 
 @Component({
   selector: 'summoner-window-last-build',
