+++ src/app/match-viewer-app/match-viewer-app.module.ts
@@ -8,10 +8,6 @@ import { TeamDisplayComponent } from './team-display/team-display.component';
 import { SummonerTileComponent } from './summoner-tile/summoner-tile.component';
 import { SharedModule } from '../shared/shared.module';
 import { SummonerTileDetailComponent } from './summoner-tile-detail/summoner-tile-detail.component';
-import { SummonerWindowFrameComponent } from './summoner-window-frame/summoner-window-frame.component';
-import { SummonerWindowContentComponent } from './summoner-window-content/summoner-window-content.component';
-import { SummonerWindowRunesComponent } from './summoner-window-runes/summoner-window-runes.component';
-import { SummonerWindowLastBuildComponent } from './summoner-window-last-build/summoner-window-last-build.component';
 
 @NgModule({
   declarations: [
@@ -19,11 +15,7 @@ import { SummonerWindowLastBuildComponent } from './summoner-window-last-build/s
     MatchViewerLayoutComponent,
     MatchDisplayComponent,
     SummonerTileComponent,
-    SummonerTileDetailComponent,
-    SummonerWindowFrameComponent,
-    SummonerWindowContentComponent,
-    SummonerWindowRunesComponent,
-    SummonerWindowLastBuildComponent
+    SummonerTileDetailComponent
   ],
   imports: [
     CommonModule,
