+++ src/app/match-viewer-app/match-viewer-model.ts
@@ -22,27 +22,6 @@ export class SummonerDetail {
     summonerRank: string;
 }
 
-export class BuildItem {
-    name: string;
-    icon: string;
-    description: string;
-    cost: number;
-}
-
-export enum RunePath {
-    Domination,
-    Precision,
-    Inspiration,
-    Sorcery,
-    Resolve
-}
-
-export class Rune {
-    name: string;
-    description: string;
-    path: RunePath;
-}
-
 export class Champion {
   championName: string;
   championIcon: string;
